pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "statcraft"
version = "0.1.0"
description = "Second-level neuroimaging analysis tool for group-level comparisons and statistical inference"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "StatCraft Contributors"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
]
keywords = ["neuroimaging", "fMRI", "PET", "second-level", "GLM", "BIDS", "nilearn"]

dependencies = [
    "nilearn>=0.10.0",
    "nibabel>=4.0.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "pybids>=0.15.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
statcraft = "statcraft.cli:main"

[project.urls]
Homepage = "https://github.com/arovai/StatCraft"
Documentation = "https://github.com/arovai/StatCraft#readme"
Repository = "https://github.com/arovai/StatCraft"
Issues = "https://github.com/arovai/StatCraft/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
statcraft = ["templates/*.html", "templates/*.css"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
